[surface model] Use long in SurfHash multimap
* use long in SurfHash multimap

Added CI target for the surface model (not in Jenkins yet)

* Enable testRaytracing testing in the CPU-based builds

* Added TEST_BENCHMARK target

* Added CI target for the surface model (not in Jenkins yet)

use radius and normal for hashing cylindrical surfaces

* use radius and normal for hashing cylindrical surfaces

+++ VecGeom/surfaces/base/CpuTypes.h
@@ -21,7 +21,7 @@ template <typename Real_t>
 struct CPUsurfData {
   using VecInt_t       = std::vector<int>;
   using VecChar_t      = std::vector<char>;
-  using MultimapInt_t  = std::multimap<int, int>;
+  using MultimapInt_t  = std::multimap<long, int>;
   using SurfData_t     = SurfData<Real_t>;
   using CylData_t      = CylData<Real_t>;
   using ConeData_t     = ConeData<Real_t>;
